Merge in AssetLibrary support for iOS 7.x
Currently for the Podfile there are two versions, one using
PhotosLibrary and one AssetLibrary. The AssetLibrary version does not
produce high-quality thumbs, so for clients needing iOS 7 support this
means using the previous version and therefore all versions have to
suffer low-quality thumbs. Makes more sense just to support the legacy
AssetLibrary code from the previous version.  I took that code and
merged into into the current base so it’s supported side-by-side and
added a flag to the image controller that checks the Foundation
version. If the PhotoLibrary is available, it uses that, otherwise it
falls back to the AssetLibrary.  Couple notes: The delegate methods now
return either PHAssets or ALAssets depending on which OS is being run
(8 or 7, respectively).  Clients will need to check the OS version or
object class. Also changed assetCollectionSubtypes to
collectionSubtypes and added a more generic enum,
QBImagePickerCollectionSubtype. A new method in QBAlbumsViewController
(assetCollectionSubtypesForSubtypes:) maps the enum to the PhotoLibrary
or AssetLibrary specific types. Also worth noting that I didn’t really
spend time trying to map all the flags to one and another, so I mapped
a few that seemed reasonable and added an “All” subtype. The All
subtype maps to the set in the current version and the set in the
previous version.