Delay asset controller layout until frame is correct and layout is complete
When presenting as a .FormSheet on iPad, the frame is not correct until just after `viewWillAppear:`. Moving the `scrollToItem:` code to viewDidAppear is too late, and viewWillAppear is too early. Dispatching to the main thread waits one run loop until the frame is update and the layout is complete without being visible to the user.